move 'port' command code around
Summary:moves code around to make a "port" service, which takes a bunch of filenames and gives you back a bunch of patches that port from docblock to flow types. This would seem more service-like if the service spawned workers instead of doing it serially, but that's an implementation detail.

The main motivation is to get `Comments_js` out of `typing/` since it uses `Str`, and to have `Server` delegate as much work as it can to services.

as a bonus to justify the churn, this now returns errors to the caller instead of dumping them in the log file.

Reviewed By: avikchaudhuri

Differential Revision: D3025574

fb-gh-sync-id: 7deaa5206fb4a476a0d01c4b22e2f57743532aa1
shipit-source-id: 7deaa5206fb4a476a0d01c4b22e2f57743532aa1