Add `all` flag to flow config
Summary:At the moment you can do stuff like `flow check --all` which
tells flow to treat all files like they have the `flow` directive. The
current test runner toggles this on and off using the `.testconfig`.

This diff instead allows the `.flowconfig` to specify whether or not to
use the `--all` behavior. In the new test runner I won't default to using
`--all`, so being able to specify it in the `.flowconfig` is useful for
supporting legacy tests

Reviewed By: avikchaudhuri

Differential Revision: D3041372

fb-gh-sync-id: 83ed1520c0d488a06cdaf58324c7ba17124dcfd3
shipit-source-id: 83ed1520c0d488a06cdaf58324c7ba17124dcfd3