Fix implementation of "object as instance" subtyping rule
Summary:The previous implementation treated any object as a subtype of an
non-structural instance type (i.e., an instance type arising from a class
declaration rather than an interface) as long as the object were structurally
similar to the instance type.

This is broken in a not-so-subtle way, but existed for so long because there
appeared to be some valid uses of this rule in the wild.

I investigated these seemingly valid uses, which fall into two buckets:

1. Using Object.create as an alternate constructor mechanism

This is a legit use case, but the existing implementation was too permissive.
The new implementation searches up the prototype chain to see the constructor
is there. If it is, then `instanceof` checks would pass at runtime, which is
sufficient to allow the object as a subtype, I believe.

2. Using objects to duck-type into non-structural instance types

This use case is not correct. Class types are nominal, which allows us to
understand `instanceof` checks correctly. For some value `x` and class `C`, `x
instanceof C` is only true if `C` is in the prototype chain of `x`. By allowing
`x` to "become" a `C` by any other means, we introduce unsoundness.

Two illustrative instances of this are:
a. Using an object to stand in for an Immutable.Record.
b. Using an object to stand in for some built-in class, like ClientRect

In the former case, an easy fix is to just create an instance of the Record
instead of lazily using an object.

The latter case is interesting, because browsers don't provide any way to
actually construct a ClientRect directly -- the constructor is private and
throws. You can only get one through DOM APIs.

I believe the correct solution to the latter case is to define an interface
type which is structurally compatible with ClientRect and the object types.
Using ClientRect can be seen as a lazy shortcut, in a sense.

Reviewed By: avikchaudhuri

Differential Revision: D3040119

fb-gh-sync-id: f07c4044eb673123673c4821e5bd1ef684ab9ce4
shipit-source-id: f07c4044eb673123673c4821e5bd1ef684ab9ce4