Fix destructuring with defaults for optional types
Summary:Before this change, destructuring a non-literal object that has optional
properties would cause "possibly undefined" errors when a default was provided.

To solve this, we need to unwrap the optional as part of the destructuring
evaluation in Flow_js. I added a selector that marks when a default is provided,
which threads the destructuring tvar through a predicate that unwraps the
optional.

Reviewed By: bhosmer

Differential Revision: D3043601

fb-gh-sync-id: 09d98851653b729ffb3b15202ac56230647dd643
shipit-source-id: 09d98851653b729ffb3b15202ac56230647dd643