[hack] Fix find_oom_in_dmesg_output function
Summary:This function assumes that ServerProcess.name is the name that will show up in
this log. But those names are unrelated - the name in the log will always be
inherited from the monitor process name. After I added the IDE process and
renamed the main process from "hh_server" to "typechecker" this function was
not working.

Reviewed By: int3

Differential Revision: D3044877

fb-gh-sync-id: 98bb5a2e562b714390b0186958e563fa037bb044
shipit-source-id: 98bb5a2e562b714390b0186958e563fa037bb044