[hack] Port hh_get_method_name, hh_ide.js version of --identify-function
Summary:I thought that we don't need it, since we have --identify-function, but the latter one is actually losing information: it drops the symbol type and position, giving you just the string.

I proposed making the IDE version of identify-function contain this data without changing the command line one, but Peter tells me that this will complicate things on Nuclide side, so adding another command. As mentioned in previous diffs, they are all up for a re-do in v2 of JSON protocol.

Reviewed By: int3

Differential Revision: D3063737

fb-gh-sync-id: 334539f8a409d37395e8552b6a634b734f5bcb13
shipit-source-id: 334539f8a409d37395e8552b6a634b734f5bcb13