`<PROJECT_ROOT>` support in .flowconfig [ignore] section
Summary:This always bugged me. Since the ignores are regexes, you couldn't
provide relative paths to ignore. With `<PROJECT_ROOT>` you can. So now you can
write

```
[ignore]
<PROJECT_ROOT>/tests/.*
```

where previously you had to write

```
[ignore]
.*/tests/.*
```

The advantage of the new version is that it doesn't ignore every directory
named tests, only the one in the project root.

This mirrors what we do in the `module.name_mapper` option

Reviewed By: mroch

Differential Revision: D3056992

fb-gh-sync-id: 6f9cafeafdb5cae70c8e0965ad98b353474ddd07
shipit-source-id: 6f9cafeafdb5cae70c8e0965ad98b353474ddd07