add --ignore-version flag to status and check
Summary:i'm tired of updating the .flowconfig to test dev versions :)

this moves the version check from the parsing of the flowconfig to when actually starting the server. this also opens the ability to add a `--version` flag to `flow init`, but I didn't do that yet because I'm not sure it should be its own flag (and its own .flowconfig section), or if we should make it part of `[options]` and therefore the `--options` flag.

Reviewed By: samwgoldman

Differential Revision: D3077339

fb-gh-sync-id: e4e3744f91e833dda3a6912f43c04cc74a7f4375
shipit-source-id: e4e3744f91e833dda3a6912f43c04cc74a7f4375