Fix false negative arising from generics and addition
Summary:
This fixes a bad interaction between polymorphic instantiation and the
addition rules. Given generic types A and B and values a:A and b:B, the
expression a+b should always evaluate to `mixed` because we don't know a
or b will be numbers or strings, and thus whether the resulting
expression will be a number or string.

This expectation follows from the existing behavior, but the rule for
ground_subtype that prunes flows with an empty lower bound prevented the
entire addition expression from becoming mixed.

This change allows (EmptyT _, AdderT _) to reach flow_addition and
handles the possible EmptyT internally. This ensures that the addition
expression always gets a lower bound (mixed) which eventually runs into
the empty instantiation of the return annotation.

Fixes #1574

Test Plan:
added tests