[hack] Small optimization in type inference: avoid immediate introduction of indirections when creating fresh unification variables
Summary:As it says on the tin. Currently, there are lots of places in Typing and Typing_arrays that
introduce a unification variable and immediately unify it against an empty unresolved,
thereby introducing an unnecessary indirection in env. I don't think there is an observable
perf improvement from avoiding this, but it leads to a cleaner environmemt that is easier
to debug when dumping using the forthcoming type checker --dump-type option.

Differential Revision: D3092787

fb-gh-sync-id: aef8766990bade7fc1bd4228848d7a16cd84bf3b
shipit-source-id: aef8766990bade7fc1bd4228848d7a16cd84bf3b