Fix check-contents w/o filename & update help text
Summary:The --help text for `check-contents` did not make it obvious what the
optional `FILE` argument was and how it worked. I updated the usage info to
make it more clear.

While doing that, I noticed that the pretty printed errors weren't showing
context for `check-contents` without `FILE`. I fixed that.

Reviewed By: mroch

Differential Revision: D3094080

fb-gh-sync-id: d5af9335fdb0bcd5393e958a1aa7dfec48f83c77
shipit-source-id: d5af9335fdb0bcd5393e958a1aa7dfec48f83c77