[hack] Add persistent debug connection
Summary:I'm trying to make sure I understand incremental mode properly since I am
modifying it for lazy decl. This provides a way to e.g. see what files are
getting incrementally rechecked at each step of the way. Can further be
generalized to print out other debug info.

This already reveals two things:
1. It looks like we are redeclaring files twice. E.g. if A changes and B
   extends A, we redeclare A, discover B needs to be redeclared as well, and
   then we redeclare A and B... the final step should just be to declare B

2. We redeclare functions even if the only things that change are the functions
  they call in their function bodies. Probably a relic from when body naming was
  done in the decl phase.

Why not just dump the data to the logfile? Mostly because the latter is not
really in a machine-parseable format. Plus we would need to gate it so that it
doesn't produce logspew and cause perf overhead in the course of normal
operation... at which point we are basically doing all the plumbing to create a
new command. Having this output in JSON will be easier to write tests against.
This approach also avoids the race conditions we currently hack around in our
integration testing -- instead of `sleep()`-ing until the logs get flushed, we
can just `select()` on the debug output.

Reviewed By: ericlippert

Differential Revision: D3108941

fb-gh-sync-id: 975234919858ddd01962ec03fd0b9b5a4ace4882
fbshipit-source-id: 975234919858ddd01962ec03fd0b9b5a4ace4882