force-recheck command
Summary:Sometimes dfind takes a while to notify that a file has changed, so when status
is called immediately after a file is changed, it is possible to get stale
results.

To solve this problem, this diff implements a command that forces recheck of a file.
Now, when force_recheck && status is called immediately after a file is changed, there are two scenarios:

* either dfind wins, in which an extra recheck is forced but this is cheap
  because we don't do much if the file hasn't changed

* or dfind loses, in which case a recheck is forced, then dfind causes an extra
  recheck, but again this is cheap because we don't do much if the file hasn't
  changed.

Thus the overhead of using this command is roughly the parsing time of a file,
which is not much.

Technically, implementing the command needs some refactoring of how client commands
are processed by the server, so that recheck can be called as part of such
processing. In particular, client commands now take and give back server
envs. This generalization also subsumes the implementation of status
--wait-for-recheck, which is thus deleted.

Reviewed By: gabelevi

Differential Revision: D3040698

fb-gh-sync-id: 4b0c92369138fc9c456cbaccd81838340142a567
fbshipit-source-id: 4b0c92369138fc9c456cbaccd81838340142a567