[hack] Return all dependencies if a decl is missing
Summary:In incremental mode, we compare old and new declarations of a class Foo to
figure out if the use sites of Foo need to be redeclared or rechecked. However, this
gets a bit tricky when a file gets deleted.

In the non-lazy-decl world, a deleted file will show up as existing in the old
files_info map, and the declaration Foo that it contains will be in the decl
heap. When we run the redecl phase, we will see that Foo exists in the old decl
heap but not the new one, realize that it is deleted, and add all its
dependencies to the set of things to be rechecked. Then we remove the old
declaration at the end of the redecl phase. However, we may need to run redecl
twice on the same file (see {D3110338} for reasons). The second time around, Foo is
neither in the old nor the new decl heaps (since we removed the old one), so
we treat it as a no-op.

However, lazy decl makes things a bit different. If Foo gets deleted, we will
obivously not see it in the new decl heap, but we might not see it in the old
one either -- laziness means that we can skip declaring it on init. So we need
to add all dependencies when we see that both the old and new decls are
missing.

Adding dependencies in more cases is obviously more conservative behavior, so
correctness shouldn't be impacted. One concern is that my analysis above may be
inaccurate, and we may be rechecking more than we were before. As a sanity
check I dumped out the list of rechecked files (using {D3108941}) when rebasing
over 400 revisions and verified it was the same before and after this change.
(400 revs is kind of small but the fact that larger updates tend to get broken
up into batches makes doing comparisons more difficult.)

Reviewed By: ericlippert

Differential Revision: D3110281

fb-gh-sync-id: 926a0ff3fb6af000fea2500cfff91c6760edbfe0
fbshipit-source-id: 926a0ff3fb6af000fea2500cfff91c6760edbfe0