[Easy][Docs] Fix object example that had error for wrong reason
Summary:This bit of documentation is making the point that Flow is a bit
unsound with objects that are built up over time and doesn't really know for
sure when properties have been initialized. Nevertheless, Flow does know the
types of the properties.

The "nevertheless" part of the example had an `undefined ~> string` error
because we forgot to return a value from the function. D'oh.

Fixes https://github.com/facebook/flow/issues/1344

Reviewed By: samwgoldman

Differential Revision: D3114251

fb-gh-sync-id: 451588fba181782207bf5e68d1d4b080984893cd
fbshipit-source-id: 451588fba181782207bf5e68d1d4b080984893cd