[travis] use fake compiler name to get a separate cache for each job
Summary:currently our travis caches step on each other. each job installs dependencies into a different folder so they're all disjoint, but it currently takes 4 builds (and is super racy) to get all 4 jobs' dependencies cached. plus, each job has to download the builds for all the other jobs, which is currently ~400MB and takes a non-trivial amount of time (about 90 seconds).

this diff attempts to use a trick suggested in https://github.com/travis-ci/travis-ci/issues/4393 whereby we give Travis a fake "compiler" setting, which gets used in the cache "slug" (the tarball's filename) but which we don't rely on in our build; we still just use gcc.

this should generate a separate cache for each job.

Reviewed By: samwgoldman

Differential Revision: D3123506

fb-gh-sync-id: a896b06c0cbcf07ee32292c24499214f80e479e3
fbshipit-source-id: a896b06c0cbcf07ee32292c24499214f80e479e3