Fix build by copying header files
Summary:The open source build was broken for two reasons:

1. There were two new .c files that needed to be listed in the Makefile
2. Hack now is using .h files and ocamlbuild doesn't copy these into _build

I did a little bit of googling to see if it is easy to get ocamlbuild to understand .h files, but [it didn't seem that easy](http://stackoverflow.com/questions/19644451/how-to-make-ocamlbuild-compile-and-link-a-c-file-into-an-ocaml-project). So I took the easy way out and just updated the Makefile to simply copy the .h files into _build.
Closes https://github.com/facebook/flow/pull/1623

Differential Revision: D3135669

Pulled By: gabelevi

fb-gh-sync-id: 392daa1df114322f84e63ea28b41f63f18813eed
fbshipit-source-id: 392daa1df114322f84e63ea28b41f63f18813eed