simplify path normalization in error printing
Summary:Errors_js was doing a lot of work to normalize paths that AFAICT it didn't need to do. I think the problem was that, when `strip_root` is true, the root is stripped off in several places. By the time it gets to `relative_path`, the path is already relative. Then `relative_path` was trying to make the relative path relative to the root, but that assumes the relative path is relative to `cwd`! So, instead, `relative_path` shouldn't do anything if the path is already relative.

I also changed `stdin_file` to take a `Path.t` so we know it is absolute, so we don't need to pass in the root to normalize it later.

Reviewed By: gabelevi

Differential Revision: D3134870

fb-gh-sync-id: 39a6fdb2bbf22e25cade962c383f851d49afb9db
fbshipit-source-id: 39a6fdb2bbf22e25cade962c383f851d49afb9db