fix isolation of tests so that failures copy .out to the right place
Summary:when a test fails, we copy the `.out`, `.err` and `.diff` files back to the source dir so that they can be more easily viewed.

recently these started getting copied into the temp dir in which we copied all of the tests.

this diff changes the runner to copy the files for each test into a temp dir for each test, which better isolates them and makes the paths simpler so they can be copied.

since some tests were depending on each other (!!), I added a `cwd` option to `.testconfig` so a test can be run in one of its subdirectories. so for example, you can have `tests/foo/a` and `tests/foo/b/.flowconfig`, and set `cwd: b` to have the tests run in `b/`; `b/.flowconfig` can include `../a`.

all of the test file moves are breaking the deps between test directories.

Reviewed By: gabelevi

Differential Revision: D3138171

fb-gh-sync-id: 7b2c1eee83774e367f9eec9fbfaea74e8851d145
fbshipit-source-id: 7b2c1eee83774e367f9eec9fbfaea74e8851d145