replace Str with String in errors_js
Summary:we can't use `Str` in js_of_ocaml. this fatals the JS when trying to json-ify the errors with context.

instead, uses `String` to count the newlines and extract the nth string. this version also avoids creating an unnecessary string for every line in the file, every time that file is mentioned in an error. I didn't measure it though.

Reviewed By: gabelevi

Differential Revision: D3135323

fb-gh-sync-id: 331e93b460a86585e194b5598a4bafe91c25f6cd
fbshipit-source-id: 331e93b460a86585e194b5598a4bafe91c25f6cd