[hack] Include "from" and "mode" columns in all client logs
Summary: For tracking responsiveness, we have the numbers on the time that successful retires took (Wall Age Ms of CLIENT_CHECK_FINISH), and we can slice it by the kind of request ("mode") and source ("from"). What is this hiding though, is the number of requests that were not successful. Those are accounted for by CLIENT_BAD_EXIT (and indirectly, by CLIENT_ESTABLISH_CONNECTION_EXCEPTION, CLIENT_CONNECT_ONCE_BUSY and CLIENT_CONNECT_TO_MONITOR_TIMEOUT). But we can't slice them by same dimensions, so it's a little bit of comparing apples to oranges.

Reviewed By: int3

Differential Revision: D3144258

fb-gh-sync-id: 33e18262df50c5e4fe53d246cb6b3a13b25b47e1
fbshipit-source-id: 33e18262df50c5e4fe53d246cb6b3a13b25b47e1