[hack] Compute name declaration position in IdentifySymbolService
Summary:Quoting #10808689:

Hyperclick (in Nuclide: Command+Click when hovering over definition, or Option+Command+Enter with cursor over definition) is currently implemented as call to --identify-function followed by call to --search with name returned by the first one.

This is highly inefficient, since search is "fuzzy search" intended for user-typed queries, it tries to be smart and use heuristics. It is also 10x-20x slower than all other queries.

We could provide another call, something like --symbol-definition, that would return the same data faster without trying to be smart, but then Hyperclick will still require two roundtrips to the server. Instead, let's return this information along with the results of --identify-function.

In this diff, I add computation of this information to the service, but the only place where it is exposed for now are unit tests.

Reviewed By: ericlippert

Differential Revision: D3169433

fb-gh-sync-id: e7e82f474252dba9136e645d4fad385423096b1b
fbshipit-source-id: e7e82f474252dba9136e645d4fad385423096b1b