Add parallelism to fuzzy search
Summary:Fuzzy search index works by maintaining buckets keyed by a letter. Each bucket contains the terms where said letter is "important" - i.e is the first letter, upper case letter, or a letter just after underscore. When processing a query, it gets the bucket for the first letter of search query and iterates over all of them computing how well do they match. This doesn't work very well for some popular letters (like E), and we can distribute this work to parallel workers.

Unfortunately, the speed up is not proportional to the number of workers - in the worst cases, just serializing the entire bucket to send it to workers take 50% of time. There also seems to be a slight regression when searching for very short terms starting with unpopular letters.

Reviewed By: alexchow

Differential Revision: D3150466

fb-gh-sync-id: 21164d028a6540356a95c05f0d518f5372d59c1d
fbshipit-source-id: 21164d028a6540356a95c05f0d518f5372d59c1d