stop running GC
Summary:since the addition of sig contexts, we're not merging complete contexts anyway, and those sig contexts prune out the same things that GC does.

disabling GC means that the context stored in ContextHeap is more intact, so we can use the cached context for things like type-at-pos and dump-types, which already have to run with gc off.

Reviewed By: bhosmer

Differential Revision: D3187474

fb-gh-sync-id: 921a320d4036175cc197cf0b2b490af84cf82412
fbshipit-source-id: 921a320d4036175cc197cf0b2b490af84cf82412